requirements.txt
asgiref==3.7.2
boto3==1.28.20
botocore==1.31.20
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
Django==4.2.4
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.2.0
hvac==1.1.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.3
Pillow==10.0.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
win32-setctime==1.1.0
